refactor: extract mesh queue send runtime
