+++ meshtastic/tests/test_mesh_interface_queue_send_runtime.py
@@ -16,13 +16,12 @@ class _QueueHarness:
             collections.OrderedDict()
         )
         self.queue_status: mesh_pb2.QueueStatus | None = None
-        self.failure: BaseException | None = None
 
     def set_queue_status(self, queue_status: mesh_pb2.QueueStatus | None) -> None:
         self.queue_status = queue_status
 
 
-def test_claim_records_queue_slot_without_status() -> None:
+def test_has_free_space_returns_true_when_no_status() -> None:
     from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
 
     harness = _QueueHarness()
@@ -35,12 +34,47 @@ def test_claim_records_queue_slot_without_status() -> None:
     )
 
     assert runtime.has_free_space()
+
+
+def test_claim_does_nothing_when_no_queue_status() -> None:
+    from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
+
+    harness = _QueueHarness()
+    runtime = _QueueSendRuntime(
+        lock=harness.lock,
+        get_queue=lambda: harness.queue,
+        get_queue_status=lambda: harness.queue_status,
+        set_queue_status=harness.set_queue_status,
+        queue_wait_delay_seconds=0.0,
+    )
+
     runtime.claim()
 
-    assert list(harness.queue.items()) == [(0, False)]
+    assert len(harness.queue) == 0
 
 
-def test_pop_for_send_skips_claim_marker_and_returns_packet() -> None:
+def test_claim_decrements_free_when_status_available() -> None:
+    from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
+
+    harness = _QueueHarness()
+    runtime = _QueueSendRuntime(
+        lock=harness.lock,
+        get_queue=lambda: harness.queue,
+        get_queue_status=lambda: harness.queue_status,
+        set_queue_status=harness.set_queue_status,
+        queue_wait_delay_seconds=0.0,
+    )
+    status = mesh_pb2.QueueStatus()
+    status.free = 3
+    status.maxlen = 10
+    harness.set_queue_status(status)
+
+    runtime.claim()
+
+    assert status.free == 2
+
+
+def test_pop_for_send_returns_oldest_entry_when_no_status() -> None:
     from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
 
     harness = _QueueHarness()
@@ -53,16 +87,55 @@ def test_pop_for_send_skips_claim_marker_and_returns_packet() -> None:
     )
     packet = mesh_pb2.ToRadio()
     packet.packet.id = 123
-    harness.queue[0] = False
     harness.queue[123] = packet
 
     popped = runtime.pop_for_send()
 
     assert popped == (123, packet)
-    assert list(harness.queue.items()) == [(0, False)]
 
 
-def test_queue_status_reply_clears_matching_packet() -> None:
+def test_pop_for_send_returns_none_when_queue_empty() -> None:
+    from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
+
+    harness = _QueueHarness()
+    runtime = _QueueSendRuntime(
+        lock=harness.lock,
+        get_queue=lambda: harness.queue,
+        get_queue_status=lambda: harness.queue_status,
+        set_queue_status=harness.set_queue_status,
+        queue_wait_delay_seconds=0.0,
+    )
+
+    popped = runtime.pop_for_send()
+
+    assert popped is None
+
+
+def test_pop_for_send_obeys_free_space_limit() -> None:
+    from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
+
+    harness = _QueueHarness()
+    runtime = _QueueSendRuntime(
+        lock=harness.lock,
+        get_queue=lambda: harness.queue,
+        get_queue_status=lambda: harness.queue_status,
+        set_queue_status=harness.set_queue_status,
+        queue_wait_delay_seconds=0.0,
+    )
+    status = mesh_pb2.QueueStatus()
+    status.free = 0
+    status.maxlen = 10
+    harness.set_queue_status(status)
+    packet = mesh_pb2.ToRadio()
+    packet.packet.id = 456
+    harness.queue[456] = packet
+
+    popped = runtime.pop_for_send()
+
+    assert popped is None
+
+
+def test_correlate_queue_status_reply_removes_matching_packet() -> None:
     from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
 
     harness = _QueueHarness()
@@ -83,6 +156,25 @@ def test_queue_status_reply_clears_matching_packet() -> None:
     runtime.correlate_queue_status_reply(status)
 
     assert 456 not in harness.queue
+
+
+def test_record_queue_status_persists_status() -> None:
+    from meshtastic.mesh_interface_runtime.queue_send import _QueueSendRuntime
+
+    harness = _QueueHarness()
+    runtime = _QueueSendRuntime(
+        lock=harness.lock,
+        get_queue=lambda: harness.queue,
+        get_queue_status=lambda: harness.queue_status,
+        set_queue_status=harness.set_queue_status,
+        queue_wait_delay_seconds=0.0,
+    )
+    status = mesh_pb2.QueueStatus()
+    status.free = 4
+    status.maxlen = 10
+
+    runtime.record_queue_status(status)
+
     assert harness.queue_status == status
